Implement Smalldots Architecture
